progress bar added, fixed minor bugs

+++ app/src/main/java/com/demo/mymovies/MainActivity.java
@@ -1,20 +1,30 @@
 package com.demo.mymovies;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
+import androidx.loader.app.LoaderManager;
+import androidx.loader.content.Loader;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.DisplayMetrics;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.CompoundButton;
+import android.widget.ProgressBar;
 import android.widget.Switch;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.demo.mymovies.adapters.MovieAdapter;
 import com.demo.mymovies.data.MainViewModel;
 import com.demo.mymovies.data.Movie;
 import com.demo.mymovies.utils.JSONUtils;
@@ -22,32 +32,77 @@ import com.demo.mymovies.utils.NetworkUtils;
 
 import org.json.JSONObject;
 
+import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Locale;
 
-public class MainActivity extends AppCompatActivity {
+public class MainActivity extends AppCompatActivity implements LoaderManager.LoaderCallbacks<JSONObject> {
     private Switch switchSort;
     private RecyclerView recyclerView;
     private MovieAdapter movieAdapter;
     private TextView textViewTopRated;
     private TextView textViewPopularity;
     private MainViewModel viewModel;
+    private ProgressBar progressBarLoading;
+
+    private static final int LOADER_ID = 17;
+    private LoaderManager loaderManager;
+    private static int page = 1;
+    private static int methodOfSort;
+    private static boolean isLoading=false;
+    private static String lang;
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.main_menu,menu);
+        return super.onCreateOptionsMenu(menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        int id = item.getItemId();
+        switch (id){
+            case R.id.itemMain:
+                Intent intent = new Intent(this, MainActivity.class);
+                startActivity(intent);
+                break;
+            case R.id.itemFavourite:
+                Intent intentToFavourite = new Intent(this, FavouriteActivity.class);
+                startActivity(intentToFavourite);
+                break;
+        }
+        return super.onOptionsItemSelected(item);
+    }
+
+    private int getColumnCount(){
+        DisplayMetrics displayMetrics = new DisplayMetrics();
+        getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
+        int width = (int)(displayMetrics.widthPixels / displayMetrics.density);
+        return width/185 > 2 ? width / 185 : 2;
+    }
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        lang = Locale.getDefault().getLanguage();
+        loaderManager = LoaderManager.getInstance(this);
         viewModel = ViewModelProviders.of(this).get(MainViewModel.class);
         textViewTopRated = findViewById(R.id.textViewTopRated);
         textViewPopularity = findViewById(R.id.textViewPopularity);
         switchSort = findViewById(R.id.switchSort);
+        progressBarLoading = findViewById(R.id.progressBarLoading);
         recyclerView = findViewById(R.id.recyclerViewPosters);
-        recyclerView.setLayoutManager(new GridLayoutManager(this,2));
+        recyclerView.setLayoutManager(new GridLayoutManager(this,getColumnCount()));
         movieAdapter = new MovieAdapter();
         switchSort.setChecked(true);
         recyclerView.setAdapter(movieAdapter);
         switchSort.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
+                page = 1;
                 setMethodOfSort(isChecked);
             }
         });
@@ -64,14 +119,18 @@ public class MainActivity extends AppCompatActivity {
         movieAdapter.setOnReachEndListener(new MovieAdapter.OnReachEndListener() {
             @Override
             public void onReachEnd() {
-                Toast.makeText(MainActivity.this,"End of the list",Toast.LENGTH_SHORT).show();
+                if(!isLoading){
+                    downloadData(methodOfSort,page);
+                }
             }
         });
         LiveData<List<Movie>> moviesFromLiveData = viewModel.getMovies();
         moviesFromLiveData.observe(this, new Observer<List<Movie>>() {
             @Override
             public void onChanged(List<Movie> movies) {
-                movieAdapter.setMovies(movies);
+                if(page==1){
+                    movieAdapter.setMovies(movies);
+                }
             }
         });
     }
@@ -87,7 +146,6 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private void setMethodOfSort(boolean isTopRated){
-        int methodOfSort;
         if(isTopRated){
             textViewTopRated.setTextColor(getResources().getColor(R.color.teal_200));
             textViewPopularity.setTextColor(getResources().getColor(R.color.white));
@@ -97,17 +155,60 @@ public class MainActivity extends AppCompatActivity {
             textViewTopRated.setTextColor(getResources().getColor(R.color.white));
             methodOfSort = NetworkUtils.POPULARITY;
         }
-        downloadData(methodOfSort,1);
+        downloadData(methodOfSort,page);
     }
 
     private void downloadData(int methodOfSort,int page){
-        JSONObject jsonObject = NetworkUtils.getJSONFromNetwork(methodOfSort,1);
-        ArrayList<Movie> movies = JSONUtils.getMoviesFromJSON(jsonObject);
+//        JSONObject jsonObject = NetworkUtils.getJSONFromNetwork(methodOfSort,1);
+//        ArrayList<Movie> movies = JSONUtils.getMoviesFromJSON(jsonObject);
+//        if(movies!=null && !movies.isEmpty()){
+//            viewModel.deleteAllMovies();
+//            for(Movie movie:movies){
+//                viewModel.insertMovie(movie);
+//            }
+//        }
+        URL url = NetworkUtils.buildURL(methodOfSort,page, lang);
+        Bundle bundle = new Bundle();
+        bundle.putString("url",url.toString());
+        loaderManager.restartLoader(LOADER_ID,bundle,this);
+    }
+
+    @NonNull
+    @Override
+    public Loader<JSONObject> onCreateLoader(int id, @Nullable Bundle args) {
+        NetworkUtils.JSONLoader jsonLoader = new NetworkUtils.JSONLoader(this,args);
+        jsonLoader.setOnStartLoadingListener(new NetworkUtils.JSONLoader.OnStartLoadingListener() {
+            @Override
+            public void onStartLoading() {
+                progressBarLoading.setVisibility(View.VISIBLE);
+                isLoading = true;
+            }
+        });
+        return jsonLoader;
+    }
+
+    @Override
+    public void onLoadFinished(@NonNull Loader<JSONObject> loader, JSONObject data) {
+        ArrayList<Movie> movies = JSONUtils.getMoviesFromJSON(data);
         if(movies!=null && !movies.isEmpty()){
-            viewModel.deleteAllMovies();
+            if(page==1){
+                viewModel.deleteAllMovies();
+                movieAdapter.clear();
+            }
             for(Movie movie:movies){
                 viewModel.insertMovie(movie);
             }
+            movieAdapter.addMovies(movies);
+            page++;
         }
+        isLoading = false;
+        progressBarLoading.setVisibility(View.INVISIBLE);
+        loaderManager.destroyLoader(LOADER_ID);
+
+    }
+
+    @Override
+    public void onLoaderReset(@NonNull Loader<JSONObject> loader) {
+
     }
 }
